refactor(main-api): shared 패키지 이동
