+++ modules/main-api/src/main/java/com/whoz_in/main_api/command/shared/application/Command.java
@@ -1,4 +1,4 @@
-package com.whoz_in.main_api.shared.application.command;
+package com.whoz_in.main_api.command.shared.application;
 
 /*
 userService.login(String userId)
