+++ modules/main-api/src/main/java/com/whoz_in/main_api/command/shared/application/CommandBus.java
@@ -1,4 +1,4 @@
-package com.whoz_in.main_api.shared.application.command;
+package com.whoz_in.main_api.command.shared.application;
 
 /*
 받은 Command에 따라 적절한 CommandHandler에게 명령을 전달하는 인터페이스이다.
