+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/device_info/application/DeviceInfoGet.java
@@ -1,6 +1,6 @@
 package com.whoz_in.main_api.query.device_info.application;
 
-import com.whoz_in.main_api.shared.application.query.Query;
+import com.whoz_in.main_api.query.shared.application.Query;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
