+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/device_info/application/DeviceInfoViewer.java
@@ -1,7 +1,7 @@
 package com.whoz_in.main_api.query.device_info.application;
 
 
-import com.whoz_in.main_api.shared.application.query.Viewer;
+import com.whoz_in.main_api.query.shared.application.Viewer;
 
 public interface DeviceInfoViewer extends Viewer {
     DeviceInfo findByDeviceId(Long deviceId);
