+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/device_info/application/DevicesInfo.java
@@ -1,7 +1,7 @@
 package com.whoz_in.main_api.query.device_info.application;
 
-import com.whoz_in.main_api.shared.application.query.Response;
-import com.whoz_in.main_api.shared.application.query.View;
+import com.whoz_in.main_api.query.shared.application.Response;
+import com.whoz_in.main_api.query.shared.application.View;
 
 public final class DevicesInfo implements Response, View {
