+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/member/application/MemberAuthInfo.java
@@ -1,6 +1,6 @@
 package com.whoz_in.main_api.query.member.application;
 
-import com.whoz_in.main_api.shared.application.query.View;
+import com.whoz_in.main_api.query.shared.application.View;
 import java.util.UUID;
 
 public record MemberAuthInfo(
