+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/member/application/MemberLogin.java
@@ -1,6 +1,6 @@
 package com.whoz_in.main_api.query.member.application;
 
-import com.whoz_in.main_api.shared.application.query.Query;
+import com.whoz_in.main_api.query.shared.application.Query;
 
 public record MemberLogin(
         String loginId,
