+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/member/application/MemberViewer.java
@@ -1,6 +1,6 @@
 package com.whoz_in.main_api.query.member.application;
 
-import com.whoz_in.main_api.shared.application.query.Viewer;
+import com.whoz_in.main_api.query.shared.application.Viewer;
 import java.util.Optional;
 
 public interface MemberViewer extends Viewer {
