+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/shared/application/QueryHandler.java
@@ -1,4 +1,4 @@
-package com.whoz_in.main_api.shared.application.query;
+package com.whoz_in.main_api.query.shared.application;
 
 /*
 형변환 없이 구현체마다 다른 Response를 제공하기 위해 제네릭을 사용했습니다.
