+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/shared/application/Response.java
@@ -1,0 +1,5 @@
+package com.whoz_in.main_api.query.shared.application;
+
+public interface Response {
+
+}
