+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/shared/application/View.java
@@ -1,4 +1,4 @@
-package com.whoz_in.main_api.shared.application.query;
+package com.whoz_in.main_api.query.shared.application;
 
 /*
 View는 쿼리용 데이터베이스로부터 가져올 값을 정의하는 VO다.
