+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/shared/application/Viewer.java
@@ -1,10 +1,10 @@
-package com.whoz_in.main_api.shared.application.query;
+package com.whoz_in.main_api.query.shared.application;
 
 /**
  * 이 인터페이스를 상속한 인터페이스는 쿼리용 데이터베이스로부터 View를 가져오는 메서드를 정의한다.
  * <p>정의하는 메서드들은 무조건 View의 구현체를 반환해야 한다.
  * <p>쿼리용 데이터베이스에 접근하는 모듈은 Viewer를 상속한 인터페이스를 구현하면 된다.
- * @see com.whoz_in.main_api.shared.application.query.View
+ * @see View
 */
 public interface Viewer {
