--- modules/main-api/src/main/java/com/whoz_in/main_api/shared/application/query/Response.java
@@ -1,5 +1,0 @@
-package com.whoz_in.main_api.shared.application.query;
-
-public interface Response {
-
-}
