+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/presentation/CommandQueryController.java
@@ -1,10 +1,10 @@
 package com.whoz_in.main_api.shared.presentation;
 
-import com.whoz_in.main_api.shared.application.command.Command;
-import com.whoz_in.main_api.shared.application.command.CommandBus;
-import com.whoz_in.main_api.shared.application.query.Query;
-import com.whoz_in.main_api.shared.application.query.QueryBus;
-import com.whoz_in.main_api.shared.application.query.Response;
+import com.whoz_in.main_api.command.shared.application.Command;
+import com.whoz_in.main_api.command.shared.application.CommandBus;
+import com.whoz_in.main_api.query.shared.application.Query;
+import com.whoz_in.main_api.query.shared.application.QueryBus;
+import com.whoz_in.main_api.query.shared.application.Response;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
